Improvements to PeriodicCondition: now it can report which nodes correspond to a periodic (shared) degree of freedom. Also replaced "Vertex" with "Node" in method names--vertex is misleading in degree > 1 FEM cases.

+++ LinearElasticity.hh
@@ -1147,7 +1147,7 @@ private:
     // Note: a "DoF" here is actually vector-valued--there are actualy
     // N * m_numDoFs variables in the elastostatic equation.
     size_t m_numDoFs = 0;
-    std::vector<int> m_dofForNode;
+    std::vector<size_t> m_dofForNode;
 
     bool m_useRigidMotionConstraint;
     std::vector<Real> m_rigidMotionConstraintRHS;
